obt_api/requirements.txt
pyserial
requests
PyQt6
